# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Skill = typeof Skill[keyof typeof Skill];


export const Skill = {
  Frontend_development: 'Frontend development',
  Backend_development: 'Backend development',
  'UI/UX_design': 'UI/UX design',
  Product_design: 'Product design',
  Product_management: 'Product management',
  User_research: 'User research',
  Data_science: 'Data science',
  Machine_learning: 'Machine learning',
  Mobile_development: 'Mobile development',
  Cloud_and_DevOps: 'Cloud and DevOps',
  Hardware_and_IoT: 'Hardware and IoT',
  Game_development: 'Game development',
  Writing_and_content: 'Writing and content',
  Marketing: 'Marketing',
  Community_building: 'Community building',
} as const;

export interface Profile {
  id: string;
  name: string;
  college: string;
  bio: string;
  skills: Skill[];
  preferredRole: string;
  hoursAvailablePerWeek: number;
  /** @nullable */
  portfolioUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProfileInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  college: string;
  /** @maxLength 500 */
  bio: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  skills: Skill[];
  /**
     * @minLength 1
     * @maxLength 80
     */
  preferredRole: string;
  /**
     * @minimum 1
     * @maximum 168
     */
  hoursAvailablePerWeek: number;
  /**
     * @maxLength 300
     * @nullable
     */
  portfolioUrl: string | null;
}

export interface ErrorResponse {
  error: string;
}