# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Authenticated student profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the signed-in student's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile has not been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Supabase is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: saveMyProfile
      tags: [profile]
      summary: Create or update the signed-in student's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Invalid profile fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Supabase is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Skill:
      type: string
      enum:
        - Frontend development
        - Backend development
        - UI/UX design
        - Product design
        - Product management
        - User research
        - Data science
        - Machine learning
        - Mobile development
        - Cloud and DevOps
        - Hardware and IoT
        - Game development
        - Writing and content
        - Marketing
        - Community building
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        college:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        preferredRole:
          type: string
        hoursAvailablePerWeek:
          type: integer
        portfolioUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - college
        - bio
        - skills
        - preferredRole
        - hoursAvailablePerWeek
        - portfolioUrl
        - createdAt
        - updatedAt
    ProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        college:
          type: string
          minLength: 1
          maxLength: 160
        bio:
          type: string
          maxLength: 500
        skills:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Skill"
        preferredRole:
          type: string
          minLength: 1
          maxLength: 80
        hoursAvailablePerWeek:
          type: integer
          minimum: 1
          maximum: 168
        portfolioUrl:
          type: ["string", "null"]
          format: uri
          maxLength: 300
      required:
        - name
        - college
        - bio
        - skills
        - preferredRole
        - hoursAvailablePerWeek
        - portfolioUrl
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error