# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Skill } from './skill';

export interface Profile {
  id: string;
  name: string;
  college: string;
  bio: string;
  skills: Skill[];
  preferredRole: string;
  hoursAvailablePerWeek: number;
  /** @nullable */
  portfolioUrl: string | null;
  createdAt: Date;
  updatedAt: Date;
}