# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Skill } from './skill';

export interface ProfileInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  college: string;
  /** @maxLength 500 */
  bio: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  skills: Skill[];
  /**
     * @minLength 1
     * @maxLength 80
     */
  preferredRole: string;
  /**
     * @minimum 1
     * @maximum 168
     */
  hoursAvailablePerWeek: number;
  /**
     * @maxLength 300
     * @nullable
     */
  portfolioUrl: string | null;
}