# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Skill = typeof Skill[keyof typeof Skill];


export const Skill = {
  Frontend_development: 'Frontend development',
  Backend_development: 'Backend development',
  'UI/UX_design': 'UI/UX design',
  Product_design: 'Product design',
  Product_management: 'Product management',
  User_research: 'User research',
  Data_science: 'Data science',
  Machine_learning: 'Machine learning',
  Mobile_development: 'Mobile development',
  Cloud_and_DevOps: 'Cloud and DevOps',
  Hardware_and_IoT: 'Hardware and IoT',
  Game_development: 'Game development',
  Writing_and_content: 'Writing and content',
  Marketing: 'Marketing',
  Community_building: 'Community building',
} as const;